refactor: 서비스 코드에서 리포지토리 코드 분리
