+++ src/main/java/kr/ac/pknu/capstone/service/DataService.java
@@ -3,7 +3,9 @@ package kr.ac.pknu.capstone.service;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import kr.ac.pknu.capstone.domain.Data.Data;
+import kr.ac.pknu.capstone.domain.Data.DataRepository;
 import kr.ac.pknu.capstone.web.dto.UpdateRequestDto;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.stereotype.Service;
@@ -15,34 +17,24 @@ import java.util.stream.Collectors;
 
 
 @Service
+@RequiredArgsConstructor
 public class DataService {
 
-    @Autowired
-    ObjectMapper objectMapper;
-
-    List<Data> data;
-
-    private void readJsonData() throws IOException {
-        ClassPathResource resource = new ClassPathResource("data.json");
-        File file = resource.getFile();
-        data = objectMapper.readValue(file, new TypeReference<List<Data>>() {});
-    }
+    private final DataRepository dataRepository;
 
     public List<Data> findAll() throws IOException {
-        if(data == null) readJsonData();
-        return data;
+        return dataRepository.findAll();
     }
 
     public List<Data> find(int vCPU, int memory) throws IOException {
-        if(data == null) readJsonData();
-        return data.stream().filter(d ->
+        // TODO: 이것도 옮기기
+        return dataRepository.findAll().stream().filter(d ->
                 (vCPU == 0 || d.getVcpu().equals(vCPU)) && ( memory == 0 || d.getMemory().equals(memory))
         ).collect(Collectors.toList());
     }
 
     public void save(UpdateRequestDto updateRequestDto) throws IOException {
-        if(data == null) readJsonData();
-        data.add(updateRequestDto.toEntity());
+        dataRepository.save(updateRequestDto.toEntity());
     }
 
 }
